refect : review, comment 몇몇 수정

+++ backend/src/main/java/com/project/backend/domain/review/comment/repository/ReviewCommentRepository.java
@@ -15,5 +15,5 @@ import java.util.List;
  */
 @Repository
 public interface ReviewCommentRepository extends JpaRepository<ReviewComment, Integer> {
-    List<ReviewComment> findAllByReviewId(Long reviewId);
+    List<ReviewComment> findAllByReviewId(Integer reviewId);
 }
